Move implementation to its own file
